Fix errors in both server and extension
Add break in messageParser and change casts to long instead of int
Add an object in localStorage to map room id to user name so that
we can specify which user the data corresponds to.
Add some error handling that is unfinished.

+++ src/server/src/messaging/ScoreGenerator.java
@@ -110,11 +110,11 @@ public class ScoreGenerator {
       for(int i = 0; i < history.size(); i++) {
         // if website entry is in bad sites, -1 score for each unit of time
         if (badSites.contains(((JSONArray)history.get(i)).get(0))){
-          score -= (int)((JSONArray)history.get(i)).get(1);
+          score -= (long)((JSONArray)history.get(i)).get(1);
         }
         // if website entry is not in bad sites, +1 score for each unit of time
         else {
-          score += (int)((JSONArray)history.get(i)).get(1);
+          score += (long)((JSONArray)history.get(i)).get(1);
         }
       }
       return score;
